Get the default model for Latent Class Analysis.

Get the default model for Latent Class Analysis.

Usage

getmodel(data, model = rep("multinomial", ncol(data)), nclasses = 2L)

Arguments

data

data.frame or matrix of response.

model

Character vector with the model for each item.

nclasses

Number of latent classes.

Value

List with the following objects:

classes

The model for the probabilities of the classes.

conditionals

The model for the conditional response probabilities.

Details

getmodel generates the model for the probability of belonging to the classes and the conditional response probabilities. These models may be modified by the user to set equality constraints or to fix parameters.

References

None yet.