inst/pkgdown/includes/custom-head.html

Convert a fitted lcfa object (latent) to a lavaan object

Convert a fitted lcfa object (latent) to a lavaan object

Usage

lcfa_to_lavaan(object, ...)

Arguments

object

A fitted lcfa object from the latent package.

...

Additional arguments passed to lavaan::cfa().

Value

A lavaan object populated with estimates from the lcfa fit.

Details

This function reconstructs a lavaan S4 object from a fitted lcfa object, allowing the use of lavaan post-processing functions such as fitMeasures(), lavResiduals(), and modindices().

The conversion works by:

Building a lavaan scaffold (un-fitted) with the same model syntax, data, estimator, and options.

Injecting lcfa parameter estimates into the lavaan parameter table via plabel matching.

Rebuilding the lavaan Model object and implied moments from the updated parameter table.

Computing test statistics, log-likelihood, baseline model, and VCOV using standard lavaan internals.

Examples

if (FALSE) { # \dontrun{
library(latent)
library(lavaan)

HS.model <- ' visual  =~ x1 + x2 + x3
              textual =~ x4 + x5 + x6
              speed   =~ x7 + x8 + x9 '

fit_latent <- lcfa(model = HS.model, data = HolzingerSwineford1939,
                   estimator = "ml", std.lv = TRUE, meanstructure = TRUE)
fit_lavaan <- lcfa_to_lavaan(fit_latent)
fitMeasures(fit_lavaan)
lavResiduals(fit_lavaan)
modindices(fit_lavaan, sort = TRUE)
} # }